% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Botete}
\alias{Botete}
\title{Biometric data of the bullseye puffer}
\format{
## `Botete`
A data frame with 1397 obs. of  7 variables:
\describe{
  \item{LT}{Total length (cm)}
  \item{LS}{Standard length (cm)}
  \item{LB}{Body trunk length (cm)}
  \item{WT}{Total weight (g)}
  \item{WB}{Body trunk weight (g)}
  \item{Wfi}{Fillet weight (g)}
  \item{Fleet}{Fleet code 1: artisanal fishing and two trawling methods}
}
}
\source{
<https://github.com/Macrurido/Morefi.git>
}
\usage{
Botete
}
\description{
Biometric data of the bullseye puffer *Sphoeroides annulatus*
collected from the Sinaloa coast of the Mexican Pacific Ocean.
}
\details{
The variable 'Fleet' classifies types of fishing: artisanal (fresh market)
as 1 and trawling (frozen-thawed market) as 2.

For further information, please refer to the article by Aguirre-Villaseñor
et al. (submitted).
}
\references{
Aguirre-Villaseñor, H., Morales-Bojórquez, E., Cisneros-Mata (FISH13711).Biometric
relationships as a fisheries management tool: A case study of the bullseye puffer
(Sphoeroides annulatus. Tetraodontidae). Fisheries Research.
}
\keyword{datasets}
