% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Botete_land}
\alias{Botete_land}
\title{Records of the bullseye puffer catch landed.}
\format{
## `Botete_land`
A data frame with 5 obs. of  5 variables:
\describe{
  \item{Category}{Landed presentation.}
  \item{Landed}{Landed weigth by category (g).}
  \item{Biomass}{Converted landed category weigth to total weight (g).}
  \item{Pesos}{Value in Mexican pesos per landed category.}
  \item{Pesos_kg}{Mean value per kilogram of landed weight.}
}
}
\source{
<https://github.com/Macrurido/Morefi.git>
}
\usage{
Botete_land
}
\description{
This data set provides records of the bullseye puffer catch landed on the
Pacific coast of Mexico in 2023. It includes information on Biomass (live
weight) for each commercial landed presentation : total weight (WT), body
trunk weight (WB), and fillet weight (Wfi). The data also indicates whether
the catch was Fresh or Frozen-thawed. Values are provided in Mexican pesos
per landed category and per kilogram of landed weight (SIPESCA, 2024).
}
\details{
Category includes five commercial landed presentations:
whole fresh (WT_Fresh);  body trunk fresh (WB_Fresh),
body trunk Fresh-Thawed (WB_Fresh_Thawed); fillet fresh (Wfi_Fresh),
and the sum of total (Total).
}
\references{
Aguirre-Villaseñor, H., Morales-Bojórquez, E., & Cisneros-Mata, M. Á. (2025).
Biometric relationships as a fisheries management tool: A case study on the
bullseye puffer (*Sphoeroides annulatus*) in an artisanal fishery. Fisheries
 Research, 292, 107593.

SIPESCA. 2024. Sistema de Información de Pesca y Acuacultura – SIPESCA.
Comisión Nacional de Pesca y Acuacultura. <https://sipesca.conapesca.gob.mx>
 (accessed 7 February 2024).
}
\keyword{datasets}
