% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_fig_cs.R
\name{fn_fig_cs}
\alias{fn_fig_cs}
\title{fn_fig_cs}
\usage{
fn_fig_cs(df, opacity, tint, scale, etiquetas, modelos, i)
}
\arguments{
\item{df}{A data frame contains the following variables: independent (x) and dependent (y) variables, the fitted variable (fitt), a weighted variable (wi), and additional details including the weights (wi), fitted values (fitt), residuals (ei), and the scale.}

\item{opacity}{A numeric value for the alpha aesthetic used to control the transparency of elements in a plot.}

\item{tint}{A vector that specifies the palette colors used to color the points.}

\item{scale}{A numeric vector that defines the thresholds for coloring the points.}

\item{etiquetas}{A list indicating the labels of the axes and their unit of measurement.}

\item{modelos}{A list of variables that are used for the regression model.}

\item{i}{An integer value indicating the regression analyzed.}
}
\value{
An individual plots for fitted models with a point color scale.
}
\description{
Graph of the fitted model with a point color scale.
}
\details{
This function creates a graph displaying the fitted model (line) alongside
the observed values (points), which are colored according to a weighted color scale.
}
\examples{
\dontrun{
df <- df
p <- fn_fig_cs(df, opacity, tint, scale, etiquetas, modelos, i)
}

}
