% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_fig_fw.R
\name{fn_fig_fw}
\alias{fn_fig_fw}
\title{fn_fig_fw}
\usage{
fn_fig_fw(df, opacity, tint, scale, order, my_labeller)
}
\arguments{
\item{df}{A data frame contains the following variables: independent (x) and dependent (y) variables, the fitted variable (fitt), a weighted variable (wi), and additional details including the weights (wi), fitted values (fitt), residuals (ei), and the scale.}

\item{opacity}{A numeric value for the alpha aesthetic used to control the transparency of elements in a plot.}

\item{tint}{A vector that specifies the palette colors used to color the points.}

\item{scale}{A numeric vector that defines the thresholds for coloring the points.}

\item{order}{A vector determines the sequence of the plots.}

\item{my_labeller}{Transforms objects to labeller functions. Used internally by labeller().}
}
\value{
A customized multi-panel plot that uses a color scale for the residual points.
}
\description{
Fitted values multi-panel plot using facet_wrap()
}
\details{
The fitted values of the models for a landed presentation category were
displayed as a multi-panel plot using the `fn_fig_fw()` function. The observed
data points for each fitted relationship were categorized according to a
weighted color scale.

Since the parameters contain subscripts, the labels were customized using the
`ggplot2::as_labeller()` function and are stored in `my_labeller`.
}
\examples{
\dontrun{
 df <- df
 opacity <- 1/5  # alpha
 tint <- c("#000000","#009E73", "#56B4E9")
 X_names <- c(LT_LS, WT_WS, LT_WT)
 wi_scale <- c(0, 0.249, 0.99, 1)
 my_labeller <- my:labeler

 p <- fn_fig_fw(df, opacity, tint, scale= wi_scale,
               order=X_names, my_labeller)
}

}
\seealso{
ggplot2::scale_colour_steps

ggplot2::as_labeller

forcats::fct_relevel
}
