% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_freqw.R
\name{fn_freqw}
\alias{fn_freqw}
\title{fn_freqw}
\usage{
fn_freqw(df, breaks, right=FALSE)
}
\arguments{
\item{df}{A vector of data values.}
\item{breaks}{A vector with class intervals.}
\item{right}{Logical, indicating if the intervals should be closed on the right (and open on the left) or vice versa}
}
\value{
A table with frequencies by class interval is returned.
}
\description{
This is a methodological function to calculate the percentage frequencies of
weights by model adjusted using the robust regression approach.
}
\details{
The function calculates the relative frequency distribution using the function `fn_freq` from `Morefi` package.

<<<<<<< HEAD
For `fn_freq`, the breaks object must be defined. For this work, a sequence of breakpoints of equal distance is used. To incorporate the number of non-weighted values wi=1, a class interval **1** is added to the sequence.
=======
For `fn_freq`, the breaks object must be defined. For this work, a sequence of breakpoints of equal distance is used. To incorporate the number of non-weighted values $(w{i}=1)$, a class interval **1** is added to the sequence.
>>>>>>> 797145bb726cc480b9f31a4fec6f464f87688996
}
\seealso{
: `cut()` and `table()` functions from the R base package.

: `fn_freq` from `Morefi` package
}
\examples{
   \dontrun{
df <- c(runif(n=30, min=0, max=1),rep(1,5))
breaks  <-  c(seq(0.1,0.9, by=0.1),1)

freq <- fn_freqw(df, breaks, right=FALSE)
          }
        }
