% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_intervals.R
\name{fn_intervals}
\alias{fn_intervals}
\title{fn_intervals}
\usage{
fn_intervals(eq, xseq)
}
\arguments{
\item{xseq}{A vector with a sequence for independent variable}
\item{eq}{eq Summary of the equation fitted}
}
\value{
df a data frame with the xseq (x1), the fitted values for dependent variable (fit), the low and upper confident (IC_L and IC_U) and predicted intervals (IP_L and IP_U)
}
\description{
The function calculates a non parametric confidence and predicted intervals
using the function predFit from the package investr (version 1.4.2)
}
\seealso{
: predFit() from the package investr (version 1.4.2)
}
\examples{
   \dontrun{
     intervalos <- fn_intervals(xseq,eq)
    }
}
