% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_summary.R
\name{fn_summary}
\alias{fn_summary}
\title{fn_summary}
\usage{
fn_summary(List_Tables, modelos, catego, eq, R2, i, j)
}
\arguments{
\item{List_Tables}{A list to store summary regression results.}

\item{modelos}{A character vector containing the variables of a function.}

\item{catego}{A character vector containing the landed weight categories.}

\item{eq}{Summary of the equation fitted.}

\item{R2}{A numeric value of the adjusted robust version of the coefficient of determination.}

\item{i}{An integer value indicating the ith regression analyzed.}

\item{j}{An integer value indicating the jth categorie analyzed.}
}
\value{
List_Tables with a customized table that stores the regression summary of each model.
}
\description{
The function customizes and stores the summary of each fitted regression.
}
\details{
Table T1 is a customized table to store the regression summarized by model
and landed weight categories: Fresh, Frozen-thawed (Frozen) and Total. The
95% confidence interval parameter is in parentheses. Residual standard error
(RSS) and Degrees of Freedom (DF) are also reported.
}
\examples{
\dontrun{
T1 <- fn_summary(List_Tables, modelos, catego, eq, R2, i, j)
}

}
\seealso{
lmrob() and nlrob() from robustbase (version 0.99-0)
}
