% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_Wlive.R
\name{fn_Wlive}
\alias{fn_Wlive}
\title{fn_Wlive}
\usage{
fn_Wlive(List_Tables, mfi, List_TCCT, Biomass)
}
\arguments{
\item{List_Tables}{A list with summary regression results.}

\item{mfi}{A character vector to define the models to be used.}

\item{List_TCCT}{A list with ARSS results.}

\item{Biomass}{A data frame with official catches.}
}
\value{
A data frame containing official catch data and estimated live weights.
}
\description{
Estimation of live weight based on landed categories.
}
\details{
The live weight, which is the total weight of an organism, is estimated based
on the weights of different landing categories, such as eviscerated weight and
fillet weight. These estimates are derived using regression parameters that
relate total weight to the weight of each landing category.
}
\examples{
 \dontrun{
 wlive <- fn_Wlive(List_Tables, mfi, List_TCCT, Biomass)
 }
}
