% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_xseq.R
\name{fn_xseq}
\alias{fn_xseq}
\title{fn_xseq}
\usage{
fn_xseq(df, modelos, bin, i)
}
\arguments{
\item{df}{A data frame includes the morphological variables.}

\item{modelos}{A list defining the numerical column variables for the model.}

\item{bin}{A vector contains the step size for each independent variable in every model.}

\item{i}{An integer value indicating the regression to be analyzed.}
}
\value{
xseq A one-column data frame with the sequence selected.
}
\description{
The function creates a data frame containing a sequence for the selected
independent variable based on the model. The lower end of the range is
determined by rounding down the minimum value to the nearest multiple of the
specified "bin" step size, while the upper end is determined by rounding up
the maximum value.
}
\details{
The data frame should have the same column name as the dependent variable
used in the fitted model.
}
\examples{
# One model: x_y

\dontrun{
x <- sample(8:44, 30, replace = TRUE)
y <- sample(20:100, 30, replace = TRUE)
df <- as.data.frame(cbind(x,y))
modelos <- list (x_y = c(1,2))
i <- 1
bin <- c(5)
xseq <- fn_xseq(df, modelos, bin, i)
}
# Two models: x_y and x_z
\dontrun{
z <- sample(100:1000, 30, replace = TRUE)
df <- cbind(x,y,z)
modelos <- list (x_y = c(1,2), y_z = c(2,3))
bin <- c(2, 100)
i <- 1
xseq1 <- fn_xseq(df, modelos, bin, i)
i <- 2
xseq2 <- fn_xseq(df, modelos, bin, i)
}

}
