% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ARSS.R
\name{fn_ARSS}
\alias{fn_ARSS}
\title{fn_ARSS}
\usage{
fn_ARSS(List_TCCT, i, alfa)
}
\arguments{
\item{List_TCCT}{A list with fitted regression results}
\item{i}{An integer value indicating the ith regression analyzed}
\item{alfa}{A numerical value that defines the significance level. If this is not set, the default number is 0.05}
}
\value{
List_TCCT calculations for ARSS are stored here
}
\description{
The function fn_ARSS perform the Coincident Curves Test, to determine if there
are significant differences between the fitted curves for each database.It is
based on the Analysis of the Residual Sum of Squares (ARSS) (Chen et al. 1992)
}
\details{
The residual sum of squares (RSS) and the degrees of freedom (DF) for each
fitted regression are previously stored in the List_TCCT list. For each
regression, the calculations are stored in a data frame T1, which is stored
iteratively using a loop for in a list T.

Inside the function, the RSS and DF for the joined sample are calculated to
perform the F test for two tails alpha/2. The decision criteria is performed:
"*" if the p-value is less than or equal to alpha or "NS" if the p-value is
greater than alpha.


The function requires defining:
}
\seealso{
: FDist from the package stats (version 3.6.2)
}
\examples{
   \dontrun{
     ARSS <- fn_ARSS(List_ARSS, i,  alfa= 0.05)
    }
}
