% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_fig_w.R
\name{fn_fig_w}
\alias{fn_fig_w}
\title{fn_fig_w}
\usage{
fn_fig_w(df, opacity, tint, my_labeller, order)
}
\arguments{
\item{df}{A data frame contains the following variables: independent (x) and dependent (y) variables, the fitted variable (fitt), a weighted variable (wi), and additional details including the weights (wi), fitted values (fitt), residuals (ei), and the scale.}

\item{opacity}{A numeric value for the alpha aesthetic used to control the transparency of elements in a plot.}

\item{tint}{A vector that specifies the palette colors used to color the points.}

\item{my_labeller}{Transforms objects to labeller functions. Used internally by labeller().}

\item{order}{A vector determines the sequence of the plots.}
}
\value{
A customized multi-panel plot that uses a color scale for the residual points.
}
\description{
Plot of residual structures classified by a Weighted Scale.
}
\details{
The residual structure was analyzed by graphing residuals against
weighted values. A custom multi-panel plot illustrates the structure of each
fitted relationship, categorized by a color-weighted scale of values.

To standardize the x-axis units across different fitted models, a standardized
residual distance, such as studentized residuals, can be used due to the
varying dimensions of the variables.

 Since the parameters contain subscripts, the labels were customized using the
`ggplot2::as_labeller()` function and are stored in `my_labeller`.
}
\examples{
 \dontrun{

 p <- fn_fig_w(df, opacity,tint, my_labeller,
               order= X_names)
}

}
\seealso{
ggplot2::as_labeller()

ggplot2::facet_wrap

forcats::fct_relevel
}
