% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_figs.R
\name{fn_figs}
\alias{fn_figs}
\title{fn_figs}
\usage{
fn_figs(datos, CIPI, i, etiquetas, modelos)
}
\arguments{
\item{datos}{A data frame containing the observed values of: independent variable (x); dependent (y) and adjusted (yc) variable; and Fleet (as factor).}
\item{CIPI}{A data frame containing the values of sequence of independent variable (xseq); fitted values, confidence and predicted interval (lower and upper values).}
\item{i}{An integer value indicating the regression analyzed.}
\item{etiquetas}{ A list indicating the labels of the axes and their unit of measurement.}
\item{modelos}{A list of variables that are used for the regression model.}
}
\value{
Fig a customized scatter plot
}
\description{
The function creates a scatter plot with the observed values (points),
fitted regression (solid line) and its confidence interval (shaded area)
}
\details{
The function use the theme_papers() to customize the appearance of the plot
and stored in the object Fig


The Inside the function, the mean (Ymean), lower (Ymin), and upper (Ymax) CI95%
are save in a table (Table_Y)
}
\seealso{
: theme_papers(): customized for the present work

: ggplot2 (version 3.4.4)
}
\examples{
   \dontrun{
    FY <- fn_figs(datos, CIPI, i, etiquetas, modelos)
    }
}
