% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_freq.R
\name{fn_freq}
\alias{fn_freq}
\title{fn_freq}
\usage{
fn_freq(df, breaks, right = FALSE)
}
\arguments{
\item{df}{A vector of data values.}

\item{breaks}{A vector with class intervals.}

\item{right}{Logical, indicating if the intervals should be closed on the right (and open on the left) or vice versa.}
}
\value{
A table with frequencies by class interval is returned.
}
\description{
The function calculates a frequency distribution of data using the functions:
`cut()` and `table()` from the R base.
}
\details{
Inside the function:
The non-cumulative absolute frequency is calculated for each user-defined
class interval within the `breaks` object.

The `breaks` object must be defined outside the function by the user.
}
\examples{
# In the df object, 30 random values between 0.1 and 0.9 are generated.
# The breaks object is a sequence of equal distances.
\dontrun{
df <- runif(n=30, min=0, max=1)
breaks <- seq(0, 1, by= 0.1)
freq <- fn_freq(df, breaks, right= FALSE)
}

}
\references{
Yau C (2024, deck 25). Frequency Distribution of Quantitative
Data. R Tutorial An R Introduction to Statistics.
https://www.r-tutor.com/elementary-statistics/quantitative-data/frequency-distribution-quantitative-data
}
\seealso{
`cut()` and `table()` from the R base package.
}
