% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_freq.R
\name{fn_freq}
\alias{fn_freq}
\title{fn_freq}
\usage{
fn_freq(df, breaks, right=FALSE)
}
\arguments{
\item{df}{A vector of data values.}
\item{breaks}{A vector with class intervals.}
\item{right}{Logical, indicating if the intervals should be closed on the right (and open on the left) or vice versa}
}
\value{
A table with frequencies by class interval is returned.
}
\description{
The function calculates a frequency distribution of data using the functions:
`cut()` and `table()` from the R base.
}
\details{
Inside the function:
<<<<<<< HEAD
The non-cumulative absolute frequency is calculated for each user defined class
interval within the `breaks` object.

The `breaks` object must be defined outside the function by the user. For example
a sequence of breakpoints of equal distance could be used.
=======
The non-cumulative absolute frequency is calculated for each user defined class interval within the `breaks` object.

The `breaks` object must be defined outside the function by the user. For example a sequence of breakpoints of equal distance could be used.
>>>>>>> 797145bb726cc480b9f31a4fec6f464f87688996
}
\seealso{
: `cut()` and `table()` functions from the R base package.
}
\examples{
   \dontrun{
    freq <- fn_freq(df, breaks, right=FALSE)
    }
}
