% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_xseq.R
\name{fn_xseq}
\alias{fn_xseq}
\title{fn_xseq}
\usage{
fn_xseq(modelos, List_x_seq, i)
}
\arguments{
\item{modelos}{A list that defines the variables that will be taken into account for the model.}

\item{List_x_seq}{A list to store the sequence of the independent variable considered for the ith fitted model.}

\item{i}{An integer value indicating the regression to be analyzed.}
}
\value{
xseq A one-column data frame with the sequence selected.
}
\description{
The function generates a data frame with a sequence for independent variable
(including minimum and maximum values) and selects it according to the model.
}
\details{
The data frame should have the same column name as the dependent variable
used in the fitted model.

Inside the function, the appropriate sequence is selected in function of model
 the mean (Ymean), lower (Ymin), and upper (Ymax) CI95% are stored in a table (Table_Y).
}
\examples{
\dontrun{
xseq <- fn_xseq(modelos,List_x_seq, i)
}

}
